consolidate weightless and weighted files
