+++ ZDD/jl_zdd/zdd.jl
@@ -1,13 +1,3 @@
-using DataStructures
-using LightGraphs
-using Statistics
-import Base: Dict
-
-include("node.jl")
-include("grid.jl")
-include("frontier.jl")
-include("edge_ordering.jl")
-
 struct ZDD_Node
     zero::Int
     one::Int
@@ -16,7 +6,7 @@ end
 mutable struct ZDD{N<:Node, S<:SimpleGraph}
     graph::Vector{ZDD_Node}
     nodes::Dict{UInt64, Int64}
-    # nodes_complete::Dict{N, Int64}      # used only when viz = True
+    # nodes_complete::Dict{N, Int64}    # used only when viz = True
     base_graph::S
     root::N
     paths::Int
@@ -44,28 +34,7 @@ function ZDD(g::SimpleGraph, root::Node; viz::Bool=false)::ZDD
 
     base_graph = g
     paths = 0
-    return ZDD(graph, nodes,base_graph, root, paths, viz)
-end
-
-# these need to be included only after the ZDD struct is defined
-include("count_enumerate.jl")
-include("visualization.jl")
-
-function num_edges(zdd::ZDD)
-    length(zdd.edges) + length(zdd.edge_multiplicity)
-end
-
-function copy_to_vec!(vec::Vector{ForbiddenPair}, set::Set{ForbiddenPair})
-    for item in set
-        push!(vec, item)
-    end
-end
-
-function reusable_unique!(vec::Vector{ForbiddenPair}, set::Set{ForbiddenPair})
-    union!(set, vec)
-    empty!(vec)
-    copy_to_vec!(vec, set)
-    empty!(set)
+    return ZDD(graph, nodes, base_graph, root, paths, viz)
 end
 
 function construct_zdd(g::SimpleGraph,
@@ -130,88 +99,17 @@ function construct_zdd(g::SimpleGraph,
     return zdd
 end
 
-
-function make_new_node(g::SimpleGraph,
-                       g_edges::Array{NodeEdge,1},
-                       k::Int,
-                       n::Node,
-                       i::Int,
-                       x::Int8,
-                       d::Int,
-                       frontiers::Array{Set{UInt8}, 1},
-                       lower_bound::Int32,
-                       upper_bound::Int32,
-                       zero_terminal::Node,
-                       one_terminal::Node,
-                       fp_container::Vector{ForbiddenPair},
-                       rm_container::Vector{ForbiddenPair},
-                       lower_vs::Vector{UInt8},
-                       recycler::Stack{Node}
-                       )
-    """
-    """
-    u = g_edges[i].edge₁
-    v = g_edges[i].edge₂
-
-    n′ = custom_deepcopy(n, recycler, x)
-
-    prev_frontier, curr_frontier = frontiers[i], frontiers[i+1]
-
-    Cᵤ, Cᵥ = components(u, v, n′)
-
-    if x == 1
-        connect_components!(n′, Cᵤ, Cᵥ)
-        @inbounds if n′.comp_weights[max(Cᵤ, Cᵥ)] > upper_bound # --> 0 if new connected component is too big
-            push!(recycler, n′)
-            return zero_terminal
-        end
-        if Cᵤ != Cᵥ && ForbiddenPair(min(Cᵤ, Cᵥ), max(Cᵤ, Cᵥ)) in n′.fps
-            push!(recycler, n′)
-            return zero_terminal
-        else
-            replace_components_with_union!(n′, Cᵤ, Cᵥ, fp_container, rm_container)
-        end
-    else
-        if Cᵤ == Cᵥ
-            push!(recycler, n′)
-            return zero_terminal
-        else
-            push!(n′.fps, ForbiddenPair(min(Cᵤ, Cᵥ), max(Cᵤ, Cᵥ)))
-        end
+function copy_to_vec!(vec::Vector{ForbiddenPair}, set::Set{ForbiddenPair})
+    for item in set
+        push!(vec, item)
     end
+end
 
-    for a in prev_frontier
-        if a ∉ curr_frontier
-            @inbounds a_comp = n′.comp_assign[a]
-            comp_assign = @view n.comp_assign[n′.first_idx:a_comp] # indices > a_comp wouldn't be labeled as a_comp, so we can ignore them
-
-            if count(x -> x == a_comp, comp_assign) == 1
-                @inbounds if n′.comp_weights[a_comp] < lower_bound
-                    push!(recycler, n′)
-                    return zero_terminal
-                end
-                @inbounds n′.comp_weights[a_comp] = 0
-                n′.cc += 1
-                if n′.cc > k
-                    push!(recycler, n′)
-                    return zero_terminal
-                end
-            end
-            remove_vertex_from_node!(n′, a, fp_container, rm_container, lower_vs)
-        end
-    end
-
-    if i == length(g_edges)
-        if n′.cc == k
-            push!(recycler, n′)
-            return one_terminal
-        else
-            push!(recycler, n′)
-            return zero_terminal
-        end
-    end
-
-    return n′
+function reusable_unique!(vec::Vector{ForbiddenPair}, set::Set{ForbiddenPair})
+    union!(set, vec)
+    empty!(vec)
+    copy_to_vec!(vec, set)
+    empty!(set)
 end
 
 function replace_components_with_union!(
@@ -246,33 +144,13 @@ function replace_components_with_union!(
     empty!(fp_container)
 end
 
-
-function connect_components!(n::Node, Cᵤ::UInt8, Cᵥ::UInt8)
-    """
-    If the two components are different, remove the smaller component from n.comp, and update n.comp_assign. Adjust n.comp_weights to remove the smaller-indexed component and add its weight into the larger one.
-    """
-    assignment = max(Cᵤ, Cᵥ)
-    to_change = min(Cᵤ, Cᵥ)
-    if Cᵤ != Cᵥ
-        map!(val -> val == to_change ? assignment : val, n.comp_assign, n.comp_assign)
-        @inbounds n.comp_weights[assignment] += n.comp_weights[to_change]
-        @inbounds n.comp_weights[to_change] = 0
-    end
-end
-
-function components(u::UInt8, v::UInt8, node::Node)::Tuple{UInt8, UInt8}
-    """ Returns Cᵤ and Cᵥ which are the sets in `components` that contain
-        vertices `u` and `v` respectively.
-    """
-    @inbounds return node.comp_assign[u], node.comp_assign[v]
-end
-
 function remove_vertex_from_node!(node::Node, vertex::UInt8, fp_container::Vector{ForbiddenPair},
                                   rm_container::Vector{ForbiddenPair}, lower_vs::Vector{UInt8})
     """
     """
     @inbounds vertex_comp = node.comp_assign[vertex]
     c = count(x -> x == vertex_comp, node.comp_assign)
+
     if c == 1
         @inbounds node.comp_assign[vertex] = 0
         for fp in node.fps
@@ -287,9 +165,17 @@ function remove_vertex_from_node!(node::Node, vertex::UInt8, fp_container::Vecto
         @inbounds node.comp_assign[vertex] = 0
         adjust_node!(node, vertex_comp, fp_container, rm_container, lower_vs)
     end
+
     if vertex == node.first_idx
         node.first_idx += 1
     end
+end
+
+function components(u::UInt8, v::UInt8, node::Node)::Tuple{UInt8, UInt8}
+    """ Returns Cᵤ and Cᵥ which are the sets in `components` that contain
+        vertices `u` and `v` respectively.
+    """
+    return node.comp_assign[u], node.comp_assign[v]
 end
 
 function lower_vertices!(num::UInt8, arr::Vector{UInt8}, container::Vector{UInt8})
@@ -302,68 +188,6 @@ function lower_vertices!(num::UInt8, arr::Vector{UInt8}, container::Vector{UInt8
         end
     end
 end
-
-function adjust_node!(node::Node,
-                      vertex_comp::UInt8,
-                      fp_container::Vector{ForbiddenPair},
-                      rm_container::Vector{ForbiddenPair},
-                      lower_vs::Vector{UInt8})
-    """
-    """
-    # there is atleast one lower vertex number that has the higher comp
-    # number and needs to be adjusted
-    lower_vertices!(vertex_comp, node.comp_assign, lower_vs) #findall(x->x==vertex_comp, node.comp_assign)
-    new_max = maximum(lower_vs)
-
-    # change comp.assign
-    for v in lower_vs
-        node.comp_assign[v] = new_max
-    end
-
-    # change comp_weights
-    if new_max != vertex_comp
-        node.comp_weights[new_max] = node.comp_weights[vertex_comp]
-        node.comp_weights[vertex_comp] = 0
-    end
-
-    # change ForbiddenPair
-    for fp in node.fps
-        if vertex_comp == fp.comp₁
-            other = fp.comp₂
-            push!(rm_container, fp)
-            push!(fp_container, ForbiddenPair(min(new_max, other), max(new_max, other)))
-        elseif vertex_comp == fp.comp₂
-            other = fp.comp₁
-            push!(rm_container, fp)
-            push!(fp_container, ForbiddenPair(min(new_max, other), max(new_max, other)))
-        end
-    end
-    filter!(x -> x ∉ rm_container, node.fps)
-    append!(node.fps, fp_container)
-    empty!(fp_container)
-    empty!(rm_container)
-end
-
-# function add_zdd_node_and_edge!(zdd::ZDD, n′::Node, n::Node, n_idx::Int64, x::Int8)
-#     """
-#     """
-#     add_vertex!(zdd.graph)
-#     n′_idx = nv(zdd.graph)
-#     zdd.nodes[n′.hash] = n′_idx
-#
-#     if zdd.viz
-#         zdd.nodes_complete[n′] = n′_idx
-#
-#         if (n, n′) in keys(zdd.edges)
-#             push!(zdd.edge_multiplicity, (n, n′))
-#         else
-#             zdd.edges[(n, n′)] = x
-#         end
-#     end
-#
-#     # add to simple graph
-#     add_edge!(zdd.graph, n_idx, n′_idx)
-# end
 
 function add_zdd_node_and_edge!(zdd::ZDD, n′::Node, n::Node, n_idx::Int64, x::Int8)
     """
